+++ v2/inbound_encoding_handler.go
@@ -26,50 +26,6 @@ import (
 	"go.uber.org/zap/zapcore"
 )
 
-// Type is an enum of RPC types.
-type Type int
-
-const (
-	// Unary types are traditional request/response RPCs.
-	Unary Type = iota + 1
-	// Streaming types are Stream based RPCs (bidirectional messages over long
-	// lived connections).
-	Streaming
-)
-
-// TransportHandlerSpec holds either a UnaryTransportHandler or StreamTransportHandler.
-type TransportHandlerSpec struct {
-	t Type
-
-	unaryHandler  UnaryTransportHandler
-	streamHandler StreamTransportHandler
-}
-
-// MarshalLogObject implements zap.ObjectMarshaler.
-func (h TransportHandlerSpec) MarshalLogObject(enc zapcore.ObjectEncoder) error {
-	enc.AddString("rpcType", h.t.String())
-	return nil
-}
-
-// Type returns the associated handler's type.
-func (h TransportHandlerSpec) Type() Type { return h.t }
-
-// Unary returns the Unary UnaryTransportHandler or nil.
-func (h TransportHandlerSpec) Unary() UnaryTransportHandler { return h.unaryHandler }
-
-// Stream returns the Stream StreamTransportHandler or nil.
-func (h TransportHandlerSpec) Stream() StreamTransportHandler { return h.streamHandler }
-
-// NewUnaryTransportHandlerSpec returns a new TransportHandlerSpec with a UnaryTransportHandler.
-func NewUnaryTransportHandlerSpec(handler UnaryTransportHandler) TransportHandlerSpec {
-	return TransportHandlerSpec{t: Unary, unaryHandler: handler}
-}
-
-// NewStreamTransportHandlerSpec returns a new TransportHandlerSpec with a StreamTransportHandler.
-func NewStreamTransportHandlerSpec(handler StreamTransportHandler) TransportHandlerSpec {
-	return TransportHandlerSpec{t: Streaming, streamHandler: handler}
-}
-
 // EncodingHandlerSpec holds either UnaryEncodingHandler or StreamEncodingHandler.
 type EncodingHandlerSpec struct {
 	t Type
@@ -100,25 +56,6 @@ func NewStreamEncodingHandlerSpec(handler StreamEncodingHandler) EncodingHandler
 	return EncodingHandlerSpec{t: Streaming, streamHandler: handler}
 }
 
-// UnaryTransportHandler handles a single, transport-level, unary request.
-type UnaryTransportHandler interface {
-	// Handle the given request.
-	//
-	// An error may be returned in case of failures. BadRequestError must be
-	// returned for invalid requests. All other failures are treated as
-	// UnexpectedErrors.
-	Handle(context.Context, *Request, *Buffer) (*Response, *Buffer, error)
-}
-
-// StreamTransportHandler handles a stream connection request in the transport layer.
-type StreamTransportHandler interface {
-	// Handle the given stream connection. The stream will close when the function
-	// returns.
-	//
-	// An error may be returned in case of failures.
-	HandleStream(stream *ServerStream) error
-}
-
 // UnaryEncodingHandler handles a single, encoding-level, unary request.
 // An encoding handler handles a request after the request has been decoded into a concrete
 // instance specific to the procedure.
@@ -136,27 +73,9 @@ type StreamEncodingHandler interface {
 	HandleStream(stream *ServerStream) error
 }
 
-// UnaryTransportHandlerFunc is a utility for defining a UnaryTransportHandler with just a
-// function.
-type UnaryTransportHandlerFunc func(context.Context, *Request, *Buffer) (*Response, *Buffer, error)
-
-// StreamTransportHandlerFunc is a utility for defining a StreamTransportHandler with just a
-// function.
-type StreamTransportHandlerFunc func(*ServerStream) error
-
 // UnaryEncodingHandlerFunc is a utility for defining a UnaryEncodingHandler with just a
 // function.
 type UnaryEncodingHandlerFunc func(context.Context, interface{}) (interface{}, error)
-
-// Handle handles an inbound unary request.
-func (f UnaryTransportHandlerFunc) Handle(ctx context.Context, req *Request, reqBody *Buffer) (*Response, *Buffer, error) {
-	return f(ctx, req, reqBody)
-}
-
-// HandleStream handles an inbound streaming request.
-func (f StreamTransportHandlerFunc) HandleStream(stream *ServerStream) error {
-	return f(stream)
-}
 
 // Handle handles an inbound unary request.
 func (f UnaryEncodingHandlerFunc) Handle(ctx context.Context, reqBody interface{}) (interface{}, error) {
