+++ v2/inbound_encoding_handler_test.go
@@ -28,37 +28,6 @@ import (
 	"go.uber.org/zap/zapcore"
 )
 
-func TestTransportHandlerSpecLogMarshaling(t *testing.T) {
-	tests := []struct {
-		desc string
-		spec TransportHandlerSpec
-		want map[string]interface{}
-	}{
-		{
-			desc: "unary_transport",
-			spec: NewUnaryTransportHandlerSpec(UnaryTransportHandlerFunc(func(context.Context, *Request, *Buffer) (*Response, *Buffer, error) {
-				return nil, nil, nil
-			})),
-			want: map[string]interface{}{"rpcType": "Unary"},
-		},
-		{
-			desc: "stream_transport",
-			spec: NewStreamTransportHandlerSpec(StreamTransportHandlerFunc(func(*ServerStream) error {
-				return nil
-			})),
-			want: map[string]interface{}{"rpcType": "Streaming"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.desc, func(t *testing.T) {
-			enc := zapcore.NewMapObjectEncoder()
-			assert.NoError(t, tt.spec.MarshalLogObject(enc), "Unexpected error marshaling spec.")
-			assert.Equal(t, tt.want, enc.Fields, "Unexpected output from marshaling spec.")
-		})
-	}
-}
-
 func TestEncodingHandlerSpecLogMarshaling(t *testing.T) {
 	tests := []struct {
 		desc string
