+++ v2/inbound_encoding_middleware_test.go
@@ -1,0 +1,51 @@
+// Copyright (c) 2018 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package yarpc_test
+
+import (
+	"context"
+	"errors"
+	"testing"
+
+	"github.com/golang/mock/gomock"
+	"github.com/stretchr/testify/assert"
+	yarpc "go.uber.org/yarpc/v2"
+	"go.uber.org/yarpc/v2/internal/internaltesttime"
+	"go.uber.org/yarpc/v2/yarpctest"
+)
+
+func TestUnaryNopInboundEncodingMiddleware(t *testing.T) {
+	mockCtrl := gomock.NewController(t)
+	defer mockCtrl.Finish()
+
+	h := yarpctest.NewMockUnaryEncodingHandler(mockCtrl)
+	wrappedH := yarpc.ApplyUnaryInboundEncodingMiddleware(h, yarpc.NopUnaryInboundEncodingMiddleware)
+
+	ctx, cancel := context.WithTimeout(context.Background(), internaltesttime.Second)
+	defer cancel()
+	reqBuf := yarpc.NewBufferBytes([]byte{1, 2, 3})
+
+	err := errors.New("great sadness")
+	h.EXPECT().Handle(ctx, reqBuf).Return(nil, err)
+
+	_, handleErr := wrappedH.Handle(ctx, reqBuf)
+	assert.Equal(t, err, handleErr)
+}
